@@ -3,10 +3,11 @@
   "version": "1.0.0",
   "description": "walk and collage",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "start": "parcel develop index.html",
-    "build": "parcel build index.html  -d docs ",
-    "watch": "parcel watch index.html  -d docs "
+    "start": "vite",
+    "build": "vite build",
+    "watch": "vite build --watch"
   },
   "dependencies": {
     "@graph-ts/vector2": "^1.3.0",
@@ -15,18 +16,23 @@
     "@types/react": "^16.9.41",
     "@types/react-dom": "^16.9.8",
     "@types/uuid": "^8.3.0",
+    "@vitejs/plugin-react": "^4.1.0",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
     "assets": "^3.0.1",
     "browser-image-resizer": "github:ericnograles/browser-image-resizer",
     "classnames": "^2.2.6",
     "jotai": "^1.6.1",
     "morphdom": "^2.6.1",
     "parcel": "^1.12.4",
+    "partysocket": "0.0.8",
     "query-string": "^7.1.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "seedrandom": "^3.0.5",
+    "vite": "^4.4.9",
     "xml2js": "^0.4.23",
+    "y-partykit": "^0.0.6",
     "y-websocket": "^1.4.0",
     "yjs": "^13.5.31"
   },
@@ -39,6 +45,7 @@
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "parcel-bundler": "^1.12.4",
+    "partykit": "0.0.23",
     "typescript": "^3.8.2"
   },
   "repository": {
